Make it work
Make the protocol and the basic requests work enough to be testable.

The big missing piece is encoding/decoding of locations, and search,
but at least the servers communicate with each other and load balance
the table, and the clients can successfully discover the servers
who are responsible for them and get/set the metadata.